Merge replicate samples

Merge replicate samples

amp_mergereplicates(data, merge_var, merge_fun = "mean", round = NULL)

Arguments

data
(required) Data list as loaded with amp_load().

merge_var
Variable in the sample metadata that defines the sample groups

merge_fun
Name of the function used to merge the samples, fx mean or sum. (default: "mean")

round
If the read counts have not been normalised, read counts are integers, so if any decimals after merging they must be rounded either "up" or "down". Make sure this makes sense if the read counts have been normalised, as it may result in 0's, 1's, and 2's everywhere. (default: NULL)

Value

An object of class ampvis2

Examples

library(ampvis2extras)
d <- amp_load(ampvis2::example_otutable, ampvis2::example_metadata)
d$metadata

#>                SampleID     Plant       Date Year
#> 16SAMP_3893 16SAMP_3893 Aalborg E 2014-02-06 2014
#> 16SAMP_3913 16SAMP_3913 Aalborg E 2014-07-03 2014
#> 16SAMP_3941 16SAMP_3941 Aalborg E 2014-08-19 2014
#> 16SAMP_3946 16SAMP_3946 Aalborg E 2014-11-13 2014
#> 16SAMP_3953 16SAMP_3953 Aalborg W 2014-02-04 2014
#> 16SAMP_4591 16SAMP_4591 Aalborg W 2014-05-05 2014
#> 16SAMP_4597 16SAMP_4597 Aalborg W 2014-08-18 2014
#> 16SAMP_4603 16SAMP_4603 Aalborg W 2014-11-12 2014

d$metadata$group <- c("group1", "group1", "group2", "group2", "group2", "group3", "group4", "group4")
d$metadata

#>                SampleID     Plant       Date Year  group
#> 16SAMP_3893 16SAMP_3893 Aalborg E 2014-02-06 2014 group1
#> 16SAMP_3913 16SAMP_3913 Aalborg E 2014-07-03 2014 group1
#> 16SAMP_3941 16SAMP_3941 Aalborg E 2014-08-19 2014 group2
#> 16SAMP_3946 16SAMP_3946 Aalborg E 2014-11-13 2014 group2
#> 16SAMP_3953 16SAMP_3953 Aalborg W 2014-02-04 2014 group2
#> 16SAMP_4591 16SAMP_4591 Aalborg W 2014-05-05 2014 group3
#> 16SAMP_4597 16SAMP_4597 Aalborg W 2014-08-18 2014 group4
#> 16SAMP_4603 16SAMP_4603 Aalborg W 2014-11-12 2014 group4

dmerged <- amp_mergereplicates(d,
  merge_var = "group",
  merge_fun = "mean",
  round = "up"
)
dmerged

#> ampvis2 object with 3 elements. 
#> Summary of OTU table:
#>      Samples         OTUs  Total#Reads    Min#Reads    Max#Reads Median#Reads 
#>            4           10        17045         3526         5451         4034 
#>    Avg#Reads 
#>      4261.25 
#> 
#> Assigned taxonomy:
#>  Kingdom   Phylum    Class    Order   Family    Genus  Species 
#> 10(100%) 10(100%) 10(100%) 10(100%) 10(100%) 10(100%)   1(10%) 
#> 
#> Metadata variables: 4 
#>  SampleID, Plant, Date, Year